Update to 2018 edition and reorganize modules
