+++ Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "freude"
 version = "0.5.0"
+edition = "2018"
 authors = ["Superfluffy"]
 license = "MIT/Apache-2.0"
