+++ src/ode/mod.rs
@@ -7,7 +7,7 @@ pub use self::tuple_impls::*;
 pub trait Ode {
     type State: Clone;
 
-    fn differentiate_into(&mut self, &Self::State, &mut Self::State);
+    fn differentiate_into(&mut self, state: &Self::State, derivative: &mut Self::State);
 
     fn differentiate(&mut self, state: &Self::State) -> Self::State {
         let mut derivative = state.clone();
