+++ src/stepper/heun.rs
@@ -1,24 +1,39 @@
 use ndarray::Dimension;
 use ndarray::IntoNdProducer;
 
-use ode::Ode;
+use crate::ode::Ode;
 
 use super::{
-    Euler,
     Stepper,
     ZipMarker,
 };
 
-impl<T> Euler<T>
+pub struct Heun<T> {
+    dt: f64,
+    dt_2: f64,
+
+    temp: T,
+    k1: T,
+    k2: T,
+}
+
+impl<T> Heun<T>
     where T: Clone,
 {
     pub fn new(state: &T, dt: f64) -> Self {
         let temp = state.clone();
+        let k1 = state.clone();
+        let k2 = state.clone();
 
-        Euler {
+        let dt_2 = dt/2.0;
+
+        Heun {
             dt: dt,
+            dt_2: dt_2,
 
             temp: temp,
+            k1: k1,
+            k2: k2,
         }
     }
 
@@ -27,15 +42,16 @@ impl<T> Euler<T>
     }
 }
 
-impl Stepper for Euler<f64>
+impl Stepper for Heun<f64>
 {
     type State = f64;
 
     fn do_step<Sy>(&mut self, system: &mut Sy, state: &mut Self::State)
         where Sy: Ode<State=f64>,
     {
-        system.differentiate_into(state, &mut self.temp);
-        self.temp = *state + &(self.dt * &self.temp);
+        system.differentiate_into(state, &mut self.k1);
+        system.differentiate_into(&(*state + &(self.dt * &self.k1)), &mut self.k2);
+        self.temp = *state + &(self.dt_2 * &self.k1) + &(self.dt_2 * &self.k2);
         system.update_state(state, &self.temp);
     }
 
@@ -44,7 +60,7 @@ impl Stepper for Euler<f64>
     }
 }
 
-impl<D, P: ZipMarker> Stepper for Euler<P>
+impl<D, P: ZipMarker> Stepper for Heun<P>
     where P: Clone,
           D: Dimension,
           for<'a> &'a P: IntoNdProducer<Dim=D, Item=&'a f64>,
@@ -56,12 +72,18 @@ impl<D, P: ZipMarker> Stepper for Euler<P>
         where Sy: Ode<State=P>,
     {
         let dt = self.dt;
-        system.differentiate_into(state, &mut self.temp);
+        let dt_2 = self.dt_2;
 
-        azip!(mut t (&mut self.temp), s (&*state) in {
-            *t = s + dt * *t;
+        system.differentiate_into(state, &mut self.k1);
+
+        azip!(mut t (&mut self.temp), s (&*state), k1 (&self.k1) in {
+            *t = s + dt * k1
         });
+        system.differentiate_into(&self.temp, &mut self.k2);
 
+        azip!(mut t (&mut self.temp), s (&*state), k1 (&self.k1), k2 (&self.k2) in {
+            *t = s + dt_2 * ( k1 + k2 )
+        });
         system.update_state(state, &self.temp);
     }
